// frontend/src/app/core/generated/circabc/model/nameValue.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * a property and their value
 */
export interface NameValue { 
    name?: string;
    value?: string;
}